Use "origin" as the default git remote explicitly.
It's really needed in the following situation. Imagine you're currently
on non-default branch, and that branch happens to have a different
remote set. Then `git ls-remote --get-url` will get the remote of
current branch, instead of that of origin. So, in order to get any kind
of sane results, we must assume default remote is "origin" and
explicitly specify it here.