Add VCS.RemoteContains method.
This will be used to detect when branches have diverged.

The git17 implementation is untested. It's a copy of git28
implementation, with the short -r flag used instead of long --remotes
flag. I think it has a good chance of working, because it's very
similar to the Contains implementation that was identical between the
two. As long as git 1.7+ supports -r flag in the same way as git 2.8+,
it should work.

There is no implementation for Mercurial at this time because I don't
use Mercurial, and don't have any repositories to test against, so the
cost is prohibitively high. It can be implemented later when there's
more demand.