Add default branch parameter to RemoteContains.
This is a necessary breaking API change to achieve improved reliability.

Some git repositories will not have the refs/remotes/origin/HEAD ref,
see https://stackoverflow.com/questions/17639383/how-to-add-missing-origin-head-in-git-repo:

> The origin's HEAD is only fetched when you clone the repo.
> If you otherwise add the remote (e.g., by using git remote add or
> by renaming another existing remote), this ref will not exist.

Therefore, the sensible thing to do is require the caller to provide
the default branch via a parameter, similar to Contains method. In
most cases, it will be easy and inexpensive for caller to do that,
since it'll have fetched the default branch by that point.